requirements.txt
fastapi==0.76.0
redmail==0.4.0
sqlalchemy==1.4.36
python-dotenv==0.20.0
